WIP move simplifyConjunctionByAssumption to Pattern simplifier
