+++ kore/src/Kore/Step/Simplification/And.hs
@@ -198,171 +198,27 @@ makeEvaluateNonBool notSimplifier sideCondition patterns = do
         from @_ @(Condition _) substitutions
         & Substitution.normalize sideCondition
     let substitution = Pattern.substitution normalized
-        predicates :: Changed (MultiAnd (Predicate variable))
+        predicates :: MultiAnd (Predicate variable)
         predicates =
             mconcat
                 [ MultiAnd.fromPredicate (predicate unified)
                 , MultiAnd.fromPredicate (predicate normalized)
                 , foldMap (from @(Predicate _) . predicate) patterns
                 ]
-            & simplifyConjunctionByAssumption
         term =
             applyAndIdempotenceAndFindContradictions
                 (Conditional.term unified)
-    case predicates of
-        Unchanged unchanged ->
-            Pattern.withCondition term (from substitution <> from predicate)
+    let predicate =
+            MultiAnd.toPredicate predicates
+            & Predicate.setSimplified simplified
+        simplified = foldMap Predicate.simplifiedAttribute predicates
+     in Pattern.withCondition term (from substitution <> from predicate)
             & return
-          where
-            predicate =
-                MultiAnd.toPredicate unchanged
-                & Predicate.setSimplified simplified
-            simplified = foldMap Predicate.simplifiedAttribute unchanged
-        Changed changed ->
-            Pattern.withCondition term (from substitution <> from predicate)
-            & simplifyCondition sideCondition
-          where
-            predicate = MultiAnd.toPredicate changed
-
-{- | Simplify the conjunction of 'Predicate' clauses by assuming each is true.
-
-The conjunction is simplified by the identity:
-
-@
-A ∧ P(A) = A ∧ P(⊤)
-@
-
- -}
-simplifyConjunctionByAssumption
-    :: forall variable
-    .  InternalVariable variable
-    => MultiAnd (Predicate variable)
-    -> Changed (MultiAnd (Predicate variable))
-simplifyConjunctionByAssumption (Foldable.toList -> andPredicates) =
-    fmap MultiAnd.make
-    $ flip evalStateT HashMap.empty
-    $ for (sortBySize andPredicates)
-    $ \predicate' -> do
-        let original = Predicate.unwrapPredicate predicate'
-        result <- applyAssumptions original
-        assume result
-        return result
-  where
-    -- Sorting by size ensures that every clause is considered before any clause
-    -- which could contain it, because the containing clause is necessarily
-    -- larger.
-    sortBySize :: [Predicate variable] -> [Predicate variable]
-    sortBySize = sortOn (size . from)
-
-    size :: TermLike variable -> Int
-    size =
-        Recursive.fold $ \(_ :< termLikeF) ->
-            case termLikeF of
-                TermLike.EvaluatedF evaluated -> TermLike.getEvaluated evaluated
-                TermLike.DefinedF defined -> TermLike.getDefined defined
-                _ -> 1 + Foldable.sum termLikeF
-
-    assume
-        :: Predicate variable
-        -> StateT (HashMap (TermLike variable) (TermLike variable)) Changed ()
-    assume predicate1 =
-        State.modify' (assumeEqualTerms . assumePredicate)
-      where
-        assumePredicate =
-            case termLike of
-                Not_ _ notChild ->
-                    -- Infer that the predicate is \bottom.
-                    HashMap.insert notChild (mkBottom sort)
-                _ ->
-                    -- Infer that the predicate is \top.
-                    HashMap.insert termLike (mkTop sort)
-        assumeEqualTerms =
-            case retractLocalFunction termLike of
-                Just (Pair term1 term2) -> HashMap.insert term1 term2
-                _ -> id
-
-        termLike = Predicate.unwrapPredicate predicate1
-        sort = termLikeSort termLike
-
-    applyAssumptions
-        ::  TermLike variable
-        ->  StateT (HashMap (TermLike variable) (TermLike variable)) Changed
-                (Predicate variable)
-    applyAssumptions replaceIn = do
-        assumptions <- State.get
-        lift $ fmap
-            (unsafeMakePredicate assumptions replaceIn)
-            (applyAssumptionsWorker assumptions replaceIn)
-
-    unsafeMakePredicate replacements original result =
-        case makePredicate result of
-            -- TODO (ttuegel): https://github.com/kframework/kore/issues/1442
-            -- should make it impossible to have an error here.
-            Left err ->
-                (error . show . Pretty.vsep . map (either id (Pretty.indent 4)))
-                [ Left "Replacing"
-                , Right (Pretty.vsep (unparse <$> HashMap.keys replacements))
-                , Left "in"
-                , Right (unparse original)
-                , Right (Pretty.pretty err)
-                ]
-            Right p -> p
-
-    applyAssumptionsWorker
-        :: HashMap (TermLike variable) (TermLike variable)
-        -> TermLike variable
-        -> Changed (TermLike variable)
-    applyAssumptionsWorker assumptions original
-      | Just result <- HashMap.lookup original assumptions = Changed result
-
-      | HashMap.null assumptions' = Unchanged original
-
-      | otherwise =
-        traverse (applyAssumptionsWorker assumptions') replaceIn
-        & getChanged
-        -- The next line ensures that if the result is Unchanged, any allocation
-        -- performed while computing that result is collected.
-        & maybe (Unchanged original) (Changed . synthesize)
-
-      where
-        _ :< replaceIn = Recursive.project original
-
-        assumptions'
-          | Exists_ _ var _ <- original = restrictAssumptions (inject var)
-          | Forall_ _ var _ <- original = restrictAssumptions (inject var)
-          | Mu_       var _ <- original = restrictAssumptions (inject var)
-          | Nu_       var _ <- original = restrictAssumptions (inject var)
-          | otherwise = assumptions
-
-        restrictAssumptions Variable { variableName } =
-            HashMap.filterWithKey
-                (\termLike _ -> wouldNotCapture termLike)
-                assumptions
-          where
-            wouldNotCapture = not . TermLike.hasFreeVariable variableName
-
-{- | Get a local function definition from a 'TermLike'.
-
-A local function definition is a predicate that we can use to evaluate a
-function locally (based on the side conditions) when none of the functions
-global definitions (axioms) apply. We are looking for a 'TermLike' of the form
-
-@
-\equals(f(...), C(...))
-@
-
-where @f@ is a function and @C@ is a constructor. @retractLocalFunction@ will
-match an @\equals@ predicate with its arguments in either order, but the
-function pattern is always returned first in the 'Pair'.
-
- -}
-retractLocalFunction
-    :: TermLike variable
-    -> Maybe (Pair (TermLike variable))
-retractLocalFunction (Equals_ _ _ term1@(App_ symbol1 _) term2@(App_ symbol2 _))
-  | isConstructor symbol1, isFunction symbol2 = Just (Pair term2 term1)
-  | isConstructor symbol2, isFunction symbol1 = Just (Pair term1 term2)
-retractLocalFunction _ = Nothing
+--         Changed changed ->
+--             Pattern.withCondition term (from substitution <> from predicate)
+--             & simplifyCondition sideCondition
+--           where
+--             predicate = MultiAnd.toPredicate changed
 
 applyAndIdempotenceAndFindContradictions
     :: InternalVariable variable
